r1789/sofa-dev : FIX: * warnings removed      * the flags of the visual context are now integer to perform a better initialization.
 //partial sync (8/9 changes)//

+++ modules/sofa/component/topology/RegularGridTopology.cpp
@@ -43,7 +43,7 @@ void RegularGridTopology::parse(core::objectmodel::BaseObjectDescription* arg)
     float scale=1.0f;
     if (arg->getAttribute("scale")!=NULL)
     {
-        scale = atof(arg->getAttribute("scale"));
+        scale = (float)atof(arg->getAttribute("scale"));
     }
 
     this->GridTopology::parse(arg);
